+++ src/main/java/shadows/apotheosis/spawn/compat/SpawnerTOPPlugin.java
@@ -35,6 +35,7 @@ public class SpawnerTOPPlugin implements TOPCompat.Provider {
 			if (spw.redstoneControl) info.mcText(SpawnerStats.REDSTONE_CONTROL.name().withStyle(TextFormatting.DARK_GREEN));
 			if (spw.ignoresLight) info.mcText(SpawnerStats.IGNORE_LIGHT.name().withStyle(TextFormatting.DARK_GREEN));
 			if (spw.hasNoAI) info.mcText(SpawnerStats.NO_AI.name().withStyle(TextFormatting.DARK_GREEN));
+			if (spw.silent) info.mcText(SpawnerStats.SILENT.name().withStyle(TextFormatting.DARK_GREEN));
 		}
 	}