+++ src/main/java/shadows/apotheosis/spawn/enchantment/CapturingEnchant.java
@@ -3,6 +3,8 @@ package shadows.apotheosis.spawn.enchantment;
 import net.minecraft.enchantment.Enchantment;
 import net.minecraft.enchantment.EnchantmentType;
 import net.minecraft.inventory.EquipmentSlotType;
+import net.minecraft.item.ItemStack;
+import shadows.apotheosis.ench.EnchModule;
 
 public class CapturingEnchant extends Enchantment {
 
@@ -23,6 +25,11 @@ public class CapturingEnchant extends Enchantment {
 	@Override
 	public int getMaxCost(int level) {
 		return this.getMinCost(level) + 15;
+	}
+
+	@Override
+	public boolean canApplyAtEnchantingTable(ItemStack stack) {
+		return super.canApplyAtEnchantingTable(stack) || EnchModule.AXE.canEnchant(stack.getItem());
 	}
 
 }